face_id_service/face_id_service/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==4.2.16
exceptiongroup==1.2.2
fastapi==0.115.2
h11==0.14.0
idna==3.10
numpy==2.0.2
opencv-python==4.10.0.84
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.39.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
